Protest/traeAi/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
websockets==11.0.3
bcrypt==4.0.1
python-dotenv==1.0.0
aiosqlite==0.19.0